Switch from --sequences-per-group to --subsample-max-sequences
This switches from the older Augur subsampling strategy of --sequences-per-group to the newer --subsampling-max-sequences. By doing this, we can just specific once the target sequence count and don't have to play around with logic for different subtypes / time-spans. This also makes it easier to rapidly test entire set of builds by locally swapping from "3000" to "100" in just one line.