Update example data to parsed metadata/sequences
Replaces unparsed sequences (with metadata in headers) with parsed
sequences and metadata as separate files. This change allows the CI
workflow to copy example data into the data directory and run the
workflow from these subtype- and segment-specific files, bypassing the
new download and filter-by-subtype rules.

One side effect of this change is that the subtype- and segment-specific
sequences and metadata now live in the `data/` directory instead of the
`results/` directory. This change makes this workflow more consistent
with other Nextstrain workflows like Zika, etc.