Use a single metadata file for all segments
Updates the "upload" workflow to create a single metadata file from the
8 individual metadata files by moving the "add segment counts" rule from
the main phylogenetic workflow to the upload workflow. As a result, all
subtypes have segment counts in their metadata regardless of whether the
"same strains" path through the phylogenetic workflow is used or not.

This commit updates the phylogenetic workflow to use a single metadata
file for all segments and retains the conditional input logic for adding
H5 clades for specific subtypes.