Convert upload to a top-level ingest workflow
Updates the upload workflow to work as a top-level ingest workflow
through a standard Snakefile entry point. As part of this
standardization, this commit moves the script for adding segment counts
into the ingest directory and updates the README to reflect constraints
on how we need to run this workflow with the Nextstrain CLI (i.e., with
the Docker runtime).