Rename example data so CI matches README
Renames the example data files to the format recommended by the README,
so the CI workflow will run the same way that users try for their first
time and we can catch any regressions. Updates the README to use the
same `cp` command that the CI workflow uses.

Adds a note about which Snakemake target to request with the example
data. Without this specific target (which is also called by the CI
workflow), the workflow will try to download the data for all other
segments from S3 and it will fail for users without credentials.