[genome] use references as root sequence
Previously the (inlined) root-sequence was inferred from the sampled
data, which meant we had stretches of gaps¹ within the root-sequence
genome. Switching to the (genbank) references as the root sequence moves
the inference of these gaps to an internal (basal) node.

¹ These are gap characters due to missing data at the 3' and 5' ends of
segments which are represented as gaps by our usage of `augur align`,
which then become internal gaps when the segments are joined.