[genome] Don't infer gaps in alignment
Most sequences have missing 3' and 5' data (for each segment) which we
were expressing as N's. In a typical analysis (in combination with
`augur ancestral --keep-ambiguous`, as used here) such terminal gaps
would be swapped out to Ns¹  but because we are concatenating segments
they are no longer terminal.

Replacing gaps with Ns solves this issue, with the side-effect that any
true deletions will be represented as Ns.

¹ <https://docs.nextstrain.org/en/latest/guides/bioinformatics/missing-sequence-data.html#gap-characters>