Add download logic to "full genome" workflow
Refactors download/copy logic and configuration variables from the main
workflow's Snakefile into a common rules file and includes this common
file in both the main workflow and the "full genome" workflow. This
change allows the full genome workflow to download/copy the files it
needs from scratch without requiring the user to first download those
file from the main workflow.

This implementation is not ideal in that the definition and appearance
of configuration variables like the S3 bucket get moved into an even
less obvious place than the top-level Snakefile. Ideally, we would
define these variables in a standard configuration YAML file. This
commit does not introduce a YAML config file yet, in favor of making a
single small improvement to the workflow.