New genome ingest ruleset
This adds a new ingest "source", namely hardcoded source-data which
is not yet available from the existing two sources. The point of this is
to produce up-to-date whole genome builds, so we tailor the ingest
for this purpose. From internal discussion¹² we are aiming to keep
this data separate to the fauna-derived (and S3 uploaded) dataset.

¹ <https://bedfordlab.slack.com/archives/CD84ELG0N/p1715729994271819>
² <https://bedfordlab.slack.com/archives/CD84ELG0N/p1715731530506199>