ingest/bin/ncbi-virus-url: pull additional fields
Thanks to comment by @AngieHinrichs¹ which linked to an example URL that
uses the `Strain_s` field. Based on this field, I was able to guess the
fields for serotype and segment.

Keeping the `isolate` field because some records use the `isolate` for
the strain name instead of the `strain` field.

Also removes the `sequence` field since that is no longer returned by
the API.²

¹ <https://github.com/nextstrain/avian-flu/issues/37#issuecomment-2123350155>
² <https://github.com/nextstrain/ingest/issues/18>