ingest/build-configs/ncbi: Add required options for NCBI Datasets
Using NCBI taxon id `11320` for Influenza A viruses to include all
GenBank records that are then further labeled with serotypes (e.g. H5N1).

Filtering down with the `--released-after` and `--geo-location`
options for NCBI datasets to get a smaller subset of data that centered
around the H5N1 outbreak in the U.S. Based on the filter URL for NCBI
virus that was shared by @trvrb:

<https://www.ncbi.nlm.nih.gov/labs/virus/vssi/#/virus?SeqType_s=Nucleotide&EnvSample_s=include&VirusLineage_ss=Alphainfluenzavirus,%20taxid:197911&Serotype_s=H5N1%20H5N*&CollectionDate_dr=2023-12-31T00:00:00.00Z%20TO%202024-05-08T23:59:59.00Z&Region_s=North%20America>