ingest/ncbi: Add steps to join Entrez with Datasets
Adds script to fetch GenBank records with Entrez for the list of
accessions that we've downloaded from NCBI Datasets. This pulls out
extra metadata that are currently not included in the NCBI Datasets,
i.e. strain, serotype, and segment. The extra metadata is then joined
with the usual NCBI Datasets metadata using `tsv-join`.

Subsequent commits will make changes the curation pipeline to account
for the extra metadata.