Switch h5n1-cattle-outbreak to NCBI data source
Switch hard-coded strain names to match those from the NCBI ingest.
Adding an assert to make sure that we are using the NCBI data source.
This currently only supports the local ingest data since I have not
added the S3 uploads to the NCBI ingest workflow (yet!).

Includes changes to the Auspice config to point to GenBank as the
data source. When testing with the new NCBI ingest data, I got a bunch
of warnings from `augur export` about filter fields not appearing
in the tree:

> WARNING:  The filter "domestic_status" does not appear as a property on any tree nodes.
> WARNING:  The filter "h5_label_clade" does not appear as a property on any tree nodes.
> WARNING:  The filter "gisaid_clade" does not appear as a property on any tree nodes.
> WARNING:  The filter "authors" does not appear as a property on any tree nodes.
> WARNING:  The filter "originating_lab" does not appear as a property on any tree nodes.

It makes sense that domestic_status, h5_label_clade, gisaid_clade,
and originating_lab are not included because they all have "?" as values.

Fixing the `author` filter that _should_ be included.