ingest/build-configs/ncbi: Add `parse-metadata-from-strain`
The strain name contains additional metadata that can supplement the
metadata from NCBI. This expects strain names to be formatted as

For non-human hosts:

    A / <host> / <location> / <sample id> / <year>

For human hosts:

    A / <location> / <sample id> / <year>

The host and year (as date) are added to the metadata if the original
metadata field is empty. The location field is specially handled
because the strain name might just contain a more detailed geolocation
than the metadata location. In that case, the strain name location is
just appended to the existing location field in the GenBank location
format <record_location>:<strain_location>.