ingest/build-configs/ncbi: Add location for A/environment/USA/CO-UW-9084466/2024
Just realized/remembered that BioSample can contain more detailed
geolocation metadata than the GenBank record. This was the only
sequence that I could find that needed the additional location data.

Will revisit the joining of BioSample data in a future change if we
see that it can fill in a lot more data.