ingest/build-configs/ncbi: Fix annotations format
Additional comments in the annotations file need to be separated by
a space after the field value. Previously separating comments by a tab
resulted in a warning from the `merge-user-metadata` script:

> WARNING: Could not decode annotation line ...

We can make the `merge-user-metadata` script smarter about parsing
comments in the annotations file, but punting that for now.