ingest: Move Andersen Lab under build-configs/ncbi/
We are planning to run this in parallel with the NCBI ingest and
eventually merge their data. They will also share more config
params with each other than with the default fauna ingest.

Andersen Lab is using NCBI SRA data, so I think it makes sense to be
under the NCBI umbrella.