ingest/andersen-lab: Update metadata/sequence outputs to match NCBI
Match the Andersen lab's metadata columns with what we are using for
NCBI so that it's easier to join them in subsequent commits.Then the
columns are reordered in a new rule to ensure the final output metadata
columns are in the same order as NCBI's based on the config's
`metadata_columns` param.

Also using NCBI config params for the FASTA id field to ensure that we
are using the same id field for sequence IDs.