ingest/andersen-lab: Remove spaces from strain names
Saw a series of warning messages from seqkit in recent ingest run
that flagged spaces in strain names.¹ This results in the sequences
being skipped during the merging of the two data sources.

We could use the `--by-name` option for seqkit as the warning message
suggested, but this reminded me that we do not support spaces in the
strain name in our downstream phylogenetic analysis anyways. So just
remove spaces from the strain name.

¹ https://github.com/nextstrain/avian-flu/actions/runs/9456655781/job/26048985198#step:13:1469