Reduce phylogeographic trait confidence
This commit as a simple --sampling-bias-correction that inflates uncertainty in traits inference. This causes SRA tips to go from highly confident in location to not at all confident in location. For example A/CATTLE/USA/24-013021-002/2024 goes from 68% to 20% confident in Ohio location.

This _should_ cause all this tips to colored gray on the tree while leaving the deeper nodes and early GenBank tips appropriately colored (which is exactly what we wanted). However, this has revealed a bug in Auspice in which Auspice currently colors tips according to most likely state unlike the logic it has for internal branches which are grayed out as entropy in state confidence increases.

However, I think this commit is still appropriate to merge in avian-flu even if we're waiting on Auspice update to properly display.