Merge snakemake.genome into main pipeline
The myriad usages of lambda functions should go away once we switch to
using config YAMLs.

Now that we have everything in a single pipeline the
h5n1-cattle-outbreak segment builds no longer need to download the
genome tree (which they depend on), they can use it as a normal
dependency and snakemake will build it as needed.