Use config YAMLs
This abstracts out the configuration into two separate YAML files. As a
result the snakemake complexity is reduced and (hopefully) the main
interaction point can now be the YAMLs themselves.

There are a few changes to the behaviour of the h5n1-cattle-outbreak
builds:
  * For individual segment builds we now use the h5n1-cattle-outbreak
    dropped list (previously we used the H5N1 drop list)
  * For individual segment builds we now use the H5NX input data
    (previously we used the H5N1 input data)
  * We no longer remove sequences via a clock filter

There are no changes to the behaviour of the GISAID builds.

The config is generally straightforward except where parameters differ
for a genome build vs the corresponding segment builds. To avoid having
to list the same parameters out 8 times, I implemented (e.g.)
`config.traits.genome_columns` and `config.traits.columns`. This is only
observed in the h5n1-cattle-outbreak config.