Add hierarchy to config directory
This (mass-renaming) is a companion to the parent commit ("Organise
results into directories"). I chose to keep the filenames unchanged as
it's a common occurrence to be editing files of the same type (e.g.
include lists) for multiple subtypes and thus having the subtype remain
in the filename is helpful.