Define path_to_fauna in config
The current value expects fauna to be a sister directory to 'ingest'.
This is valid when we are using the docker runtime¹ as it mounts the
'ingest' folder to '/nextstrain/build' and fauna is 'nextstrain/fauna',
but it almost never true when running outside of docker as that would
require fauna to be within the 'avian-flu' repo. The docs do state that
we must run in docker, but that's an overly strict requirement.

Addition context on slack <https://bedfordlab.slack.com/archives/CD84ELG0N/p1723666776588239?thread_ts=1720569955.202459&cid=CD84ELG0N>

¹ until we add a nextstrain-pathogen.yaml file