Add nextstrain-pathogen.yaml
This results in a different mount point for the 'ingest' directory in
docker runtimes which more closely matches a typical setup using other
runtimes. This allows the relative location of fauna ('path_to_fauna')
to be the same for a typical ambient setup as for docker.

See <https://github.com/nextstrain/cli/issues/371#issuecomment-2299670219>
for more detail on mount points for the docker runtime.