ingest: Dedup by sample id within NCBI and Andersen lab data
In investigating the duplicates dropped from the joined-ncbi metadata,
I realized that these duplicates were not purely from the merge of
the two data sources.

This commit deduplicates by sample id in the upstream metadata as well.
There's no need to change the processing of sequence FASTAs at this
point because they are still matched by their respective accessions
instead of strain name.