Use per-build metadata files
The previous implementation generated a metadata file for each subtype
(i.e. before filtering) and then used that for all downstream rules.
It's more common to (also) generate a post-subsampling metadata file
and use this instead. This is both smaller and useful for debugging
(as all the data in the TSV file are used in the build). It's also
going to make generating colours for specific builds much easier,
which I'll do in subsequent commits.