Simplify metadata modifications
This shifts the modification of metadata which is specific to the
h5n1-cattle-outbreak/genome/default build to be downstream of the
subsetted-metadata file, which is much simpler to reason with.

We could consider doing something similar for the `add_h5_clade` rule,
however this wouldn't allow us to use that data as a filtering criteria,
and it seems plausible we'd want to do that one day.