WIP Change config syntax
Changes the syntax for a rule parameter for to either be a scalar or a
dictionary where they keys represent the wildcards of the associated
build, allowing '*' to represent a catch-all for a wildcard. This allows
significant simplifications throughout the workflow and config YAMLs,
largely because this format lets us express parameters for any wildcard
combination without the need to maintain different structures for
genome-specific configs, or different structures for parameters which
vary per segment vs those which don't.

The subsequent commits will make more involved changes to the snakemake
pipeline facilitated by this change, this commit represents the minimal
set of changes to keep functionality.