Add config.target_patterns
A common use case is to want intermediate files (e.g. filtered metadata)
and this is a pain when we have many build targets as you have to
list these out on the command line manually. The other approach is
to modify the snakefile targets within the Snakefile, which is my
interpretation of the recent (commented out) additions in
18958650bef5a7ffa93ccc88f56867f45f5ccdc9:

```
rule all:
    input:
        auspice_json = collect_builds()
        #sequences = expand("results/{subtype}/{segment}/sequences.fasta", segment=SEGMENTS, subtype=SUBTYPES),
        #metadata = expand("results/{subtype}/metadata.tsv", segment=SEGMENTS, subtype=SUBTYPES)
```

Allowing config overlays to set these is really nice. The above
snakefile changes can now be replaced with the following config (even
more important when we start thinking about analysis directories
separate to the pathogen repo):

```yaml
extends: gisaid.yaml
target_patterns:
  - "results/{subtype}/{segment}/sequences.fasta"
  - "results/{subtype}/metadata.tsv"
```