H5N1 D1.1 genome build
This builds off the GenoFLU metadata added in previous commits to build
a D1.1-specific build. We reuse ~all of the cattle-outbreak machinery
but at this time only build the genome trees.

Due to the SRA data having limited geographic and temporal metadata
we default to the divergence tree and hide the map panel. As of
2025-02-20 the build has 615 genomes of which 412 (2/3rds) have only
the collection year and country.