[genome builds] Implement masking
Masks out positions in the genome which have nucleotides called in <50%
of samples. This is especially important in genome builds as the
terminal ends of segments are no longer terminal and thus sparse
sequence data was resulting in artefactual partitioning of the tree.

I've left the cattle-flu builds unchanged (i.e. no masking) but we
should revisit this.