WIP: ingest: Add rules for GenoFLU
Add rules to run the vendored GenoFLU on sequences. Creates a
final_metadata.tsv that has a new column `genoflu_genotype`.

This is currently not directly runnable via `nextstrain build` because
we do not have the GenoFLU dependencies in the Nextstrain runtimes.

I was able to run this locally by installing the genoflu dependencies
in my Nextstrain conda runtime with

```
mamba install -c conda-forge -c bioconda genoflu \
    --prefix ~/.nextstrain/runtimes/conda/env/ \
    --platform osx-64
```

Then I could run the new rules for the NCBI data with
```
nextstrain build --conda ingest \
    joined-ncbi/results/final_metadata.tsv \
    --configfile build-configs/ncbi/defaults/config.yaml
```