Update GenoFLU snakemake pipeline
Combined with the previous commit we now have the ability to produce
'ingest/<data-source>/results/metadata.tsv' with GenoFLU calls.
This is config-controllable and is currently set up to run GenoFLU
for NCBI & Andersen lab but not for Fauna.

The number of threads for actually running GenoFLU has been incresed to
12 as (from my testing) each thread has low CPU & memory usage, so
setting a large number of threads (even threads >> cores) improves
performance. We should revisit what exactly is happening here.

Note that the filtering approach for fauna may not be correct as
implemented here - see <https://github.com/nextstrain/avian-flu/pull/127#issuecomment-2669102995>
however fauna is currently not run through GenoFLU.