Combine color-related rules
Both simplifies the DAG and allows GISAID builds which have GenoFLU
colorings to get a nicely ordered set of colours.

Note that H9N2 & H7N9 builds, which don't export genoflu metadata, will
get colors assigned by the colors rule (because there is a value of "Not
assigned (too divergent)" in the metadata TSV) which is then dropped by
export.