Move to separate entrypoints (Snakefiles)
This starts the process of redefining the workflow interface from a
single workflow / entrypoint /Snakefile with different configfiles to
choose behaviour to an interface of multiple Snakefiles, each with their
own default configs. Ultimately this will facilitate running the
workflows in separate working directories and workflows-as-programs
(`nextstrain run ...`)

Note that because we cannot use the same config file to build both D1.1
and B3.13 builds we use a separate workflow for each. With the move to
a more expressive config syntax in <https://github.com/nextstrain/avian-flu/pull/104>
we will drop the D1.1 workflow.

I initially chose `ncbi/` and `gisaid/` as the entrypoint/workflow names
but moved away from that as I want to decouple the workflow from the
data source. As described in the README changes here it's trivial to run
either workflow using either data input. One day we may even merge these
workflows into a single config-definable workflow, but if we ever do so
it feels a long way off.