Allow running in separate analysis dir
Preferentially resolving paths to the analysis directory allows us to
run workflows in a separate workflow and override certain files
(defaults) as needed. Combined with the nicety of automatically using a
working directories `config.yaml` this allows us to run workflows from
separate analysis directories in both a convenient and powerful way.

Note that when we need to combine path resolution with resolving the
underlying config value itself we lose the benefit of currying - e.g.
the following is clunky and can be improved in a subsequent commit:

```
 colors = lambda w: resolve_config_path(get_config('colors', 'hardcoded', w))(w),
```