New "glob" config syntax
The nature of the avian-flu workflow(s) is that different parameters
vary across different wildcards. For instance, "min-length" varies
across segments, the columns used for augur traits varies according to
subtype, and others vary according to two or three of the wildcards.
Historically this has been very hard to convey succinctly in our config
syntax, and the previous implementation used a bespoke syntax which was
inconsistent across different parameters. Additionally, with the
introduction of genome-builds we ended up implementing a number of
"genome-specific" keys, such as 'clock_filter_iqd' and
'genome_clock_filter_iqd'.

Here we introduce a completely new config syntax, referred to as "glob"
syntax. This allows each parameter to succinctly represent their
different values according to a range of wildcards, and for the same
schema used for every parameter. The specifics are detailed within the
`resolve_config_value` docstring, but the above clock-filter example is
now expressible as

```yaml
    clock_filter_iqd:
        "*/genome/*": 6
        "*/*/*": false
```

This allows significant simplifications throughout the workflow and
config YAMLs due to the expressiveness and terseness of the syntax.

We make a corresponding change to the definition of 'builds' (what
wildcard combinations the workflow will construct). The new structure is
both more flexible and more amenable to config overlays. Specifying the
segments per subtype (or set of subtypes) allows a pipeline to produce
different sets of segments (e.g. ha/na for some subtypes, all 8 for
others). Encoding as a list makes it much simpler to override via
additional configs, analogous to the lists we use for 'inputs' and
'additional_inputs'. For instance, here's a GISAID overlay which just
does what you expect:

```yaml
builds:
  - subtype: h5n1
    segment:
      - ha
      - na
    time: 2y
```

This necessitated changes to the function used to gather together all
the targets of the workflow, and we add a config-definable
'target_patterns' which allows config overlays to adjust the default
targets of a workflow. (This was in response to our previous approach
which modified the snakefile to produce different targets, e.g. see
<https://github.com/nextstrain/avian-flu/commit/18958650bef5a7ffa93ccc88f56867f45f5ccdc9>)