Improve how 'subtype_query' is defined
Brings the config structure in line with all of the other config values,
but this necessitates a more complicated usage of
`resolve_config_value`. Hopefully the added comments are instructive to
people in the future!

This also greatly improves the error handling if we forget to add the
appropriate subtype to the config, or leave out 'subtype_query'
altogether.

Inspired by <https://github.com/nextstrain/avian-flu/pull/104#discussion_r1982166632>
and <https://github.com/nextstrain/avian-flu/pull/146#discussion_r1983962848>