rules/merge_inputs: Handle decompression of single inputs
Modified from the dynamic rules in the pathogen-repo-guide.¹ 

The dynamic rules did not work in this repo because it allows different number 
of inputs per segment wildcard. There's no good way to define different rules 
per wildcard so I've opted to use if/else blocks in the shell to run either 
`augur read-file` or `augur merge` depending on the number of inputs. 

¹ <https://github.com/nextstrain/pathogen-repo-guide/blob/cb87e1ac0e6c98d08bab76b954309aeb3bae889b/phylogenetic/rules/merge_inputs.smk>