Add build-configs/ci/config.yaml
The main motivation for this is to simplify the CI build-args which will 
need to be replicated across augur, docker-base, and conda-base. 

Note the CI config uses the repo's local example_data as inputs and the paths 
are relative to the top level directory, so the workflow can only be run from 
the top level directory.