WIP [ingest] GenoFLU workflow for all-influenza ingest
This takes the output from our all-influenza curation pipeline
(pre-filtered to avian-flu subtypes) and runs GenoFLU on it. It's a
little strange to have most of the ingest steps in one location and then
the GenoFLU step here; one day we may wish to unify them but that's
quite a big task given that this (avian-flu) ingest pipeline already
exists and is being used on other data sources.